find_a_friend/requirements.txt
flask
python-dotenv
numpy
mistralai

